refactor(ui): group loose root components into feature folders (notes/chat/app)
