+++ src/ui/chat/mod.rs
@@ -8,6 +8,9 @@ mod typing_indicator;
 mod user_bubble;
 mod view;
 
+pub(crate) mod action_card;
+pub(crate) mod chat_input;
+
 pub use actions::md_to_html;
 pub use sources_accordion::NoteWebSources;
 pub use view::ChatView;
