+++ src/ui/chat/view.rs
@@ -3,12 +3,12 @@ use crate::infrastructure::audio::RecordingState;
 use crate::infrastructure::persistence::Database;
 use crate::ui::chat::actions::{load_messages_from_db, send_question};
 use crate::ui::chat::bot_bubble::BotBubble;
+use crate::ui::chat::chat_input::ChatInputBar;
 use crate::ui::chat::empty_state::ChatEmptyState;
 use crate::ui::chat::menu::ChatMenu;
 use crate::ui::chat::models::ChatMsg;
 use crate::ui::chat::typing_indicator::TypingIndicator;
 use crate::ui::chat::user_bubble::UserBubble;
-use crate::ui::chat_input::ChatInputBar;
 use crate::ui::state::View;
 use crate::ui::AppState;
 use dioxus::prelude::*;
