+++ src/ui/notes/mod.rs
@@ -8,4 +8,9 @@ mod note_actions;
 mod reminders;
 mod tags;
 
+pub(crate) mod attachment_modal;
+pub(crate) mod folder_picker;
+pub(crate) mod note_card;
+pub(crate) mod note_list;
+
 pub use detail::NoteDetail;
