+++ src/ui/notes/note_card.rs
