+++ src/ui/notes/note_list.rs
@@ -2,7 +2,7 @@ use crate::application::i18n::t;
 use crate::domain::{Note, Thread};
 use crate::infrastructure::persistence::Database;
 use crate::ui::icons::*;
-use crate::ui::note_card::NoteCard;
+use crate::ui::notes::note_card::NoteCard;
 use crate::ui::thread::ThreadCard;
 use crate::ui::AppState;
 use dioxus::prelude::*;
